requirements/base.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.2.7
django-admin-tools==0.9.1
django-debug-toolbar==3.2.2
django-domainname==0.1.1
django-phonenumber-field==6.1.0
djangorestframework==3.12.4
h11==0.13.0
idna==3.3
lxml==4.9.0
outcome==1.1.0
phonenumbers==8.12.48
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
PyVirtualDisplay==3.0
requests==2.27.1
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
